fix(sn-manager): remove automatic EVM rollback
Treat evm_key_name as a transitional migration input rather than a permanent readiness marker. Already-EVM-capable supernodes now remain on the forward update path after migration removes that field; pre-v2.6 nodes without it remain blocked.\n\nRemove the v2.5.0 rollback/download/config mutation path, preserve historical marker visibility, and make supernode startup explicitly reject unsupported active key types.\n\nVerified with sn-manager unit/vet/race tests, full supernode tests, integration tests, independent review, and a 10-validator/10-supernode EVM devnet restart rehearsal.
