+++ sn-manager/cmd/status.go
@@ -17,9 +17,8 @@ var statusCmd = &cobra.Command{
 	RunE:  runStatus,
 }
 
-// printPreflightStatus prints update-blocked / rolled-back state, if any.
-// Both markers live in the manager home directory and are written by the
-// auto-updater's EVM preflight/rollback code path (see internal/updater/).
+// printPreflightStatus prints current update-blocked state and any historical
+// rolled-back marker left by v2.6.1.
 func printPreflightStatus(home string) {
 	if data, err := os.ReadFile(updater.BlockLogPath(home)); err == nil && len(data) > 0 {
 		fmt.Println("  Update Blocked: true")
@@ -31,7 +30,7 @@ func printPreflightStatus(home string) {
 		}
 	}
 	if data, err := os.ReadFile(updater.RollbackLogPath(home)); err == nil && len(data) > 0 {
-		fmt.Println("  Rolled Back: true")
+		fmt.Println("  Historical Rollback Marker (v2.6.1): present")
 		for _, line := range splitStatusLines(string(data)) {
 			if line == "" {
 				continue
