+++ sn-manager/internal/updater/preflight_test.go
@@ -14,10 +14,10 @@ func TestDecidePreflight_Table(t *testing.T) {
 		{"no-evm chain forward", preflightInputs{false, "", "v2.5.0-testnet", "v2.6.0-testnet"}, preflightAllow},
 		{"no-evm chain already 2.6", preflightInputs{false, "", "v2.6.0-testnet", "v2.6.0-testnet"}, preflightAllow},
 		{"evm no-key forward to 2.6", preflightInputs{true, "", "v2.5.0-testnet", "v2.6.0-testnet"}, preflightBlock},
-		{"evm no-key stuck at 2.6", preflightInputs{true, "", "v2.6.0-testnet", "v2.6.0-testnet"}, preflightRollback},
-		{"evm no-key stuck target 2.6.1", preflightInputs{true, "", "v2.6.0-testnet", "v2.6.1-testnet"}, preflightRollback},
-		{"evm migrated forward to 2.6", preflightInputs{true, "evm-key", "v2.5.0-testnet", "v2.6.0-testnet"}, preflightAllow},
-		{"evm migrated at 2.6", preflightInputs{true, "evm-key", "v2.6.0-testnet", "v2.6.0-testnet"}, preflightAllow},
+		{"evm migrated config at 2.6 omits transitional key", preflightInputs{true, "", "v2.6.0-testnet", "v2.6.0-testnet"}, preflightAllow},
+		{"evm migrated config can update beyond 2.6", preflightInputs{true, "", "v2.6.0-testnet", "v2.6.1-testnet"}, preflightAllow},
+		{"evm prepared forward to 2.6", preflightInputs{true, "evm-key", "v2.5.0-testnet", "v2.6.0-testnet"}, preflightAllow},
+		{"evm prepared at 2.6", preflightInputs{true, "evm-key", "v2.6.0-testnet", "v2.6.0-testnet"}, preflightAllow},
 		{"evm no-key below threshold 2.5.1", preflightInputs{true, "", "v2.5.0-testnet", "v2.5.1-testnet"}, preflightAllow},
 		{"evm no-key below threshold 2.5.0", preflightInputs{true, "", "v2.4.5-testnet", "v2.5.0-testnet"}, preflightAllow},
 	}
